split files, including network to mps, mpo
