+++ src/mqt/yaqs/analog/mcwf.py
@@ -39,7 +39,8 @@ from ..core.methods.matrix_exponential import expm_arnoldi, expm_krylov
 if TYPE_CHECKING:
     from numpy.typing import NDArray
 
-    from ..core.data_structures.networks import MPO, MPS
+    from ..core.data_structures.mpo import MPO
+    from ..core.data_structures.mps import MPS
     from ..core.data_structures.noise_model import NoiseModel
     from ..core.data_structures.simulation_parameters import AnalogSimParams
