+++ src/mqt/yaqs/core/methods/bug.py
@@ -25,7 +25,8 @@ from .tdvp import update_left_environment, update_right_environment, update_site
 if TYPE_CHECKING:
     from numpy.typing import NDArray
 
-    from ..data_structures.networks import MPO, MPS
+    from ..data_structures.mpo import MPO
+    from ..data_structures.mps import MPS
     from ..data_structures.simulation_parameters import AnalogSimParams
