+++ src/mqt/yaqs/core/methods/scheduled_jumps.py
@@ -21,7 +21,7 @@ import opt_einsum as oe
 from ..methods.tdvp import merge_mps_tensors, split_mps_tensor
 
 if TYPE_CHECKING:
-    from ..data_structures.networks import MPS
+    from ..data_structures.mps import MPS
     from ..data_structures.noise_model import NoiseModel
     from ..data_structures.simulation_parameters import AnalogSimParams
