+++ src/mqt/yaqs/core/methods/stochastic_process.py
@@ -28,7 +28,7 @@ from ..methods.tdvp import merge_mps_tensors, split_mps_tensor
 if TYPE_CHECKING:
     from numpy.typing import NDArray
 
-    from ..data_structures.networks import MPS
+    from ..data_structures.mps import MPS
     from ..data_structures.noise_model import NoiseModel
     from ..data_structures.simulation_parameters import AnalogSimParams, StrongSimParams, WeakSimParams
