+++ src/mqt/yaqs/digital/equivalence_checker.py
@@ -21,7 +21,7 @@ from typing import TYPE_CHECKING
 
 from qiskit.converters import circuit_to_dag
 
-from ..core.data_structures.networks import MPO
+from ..core.data_structures.mpo import MPO
 from .utils.mpo_utils import iterate
 
 if TYPE_CHECKING:
