+++ tests/core/data_structures/test_mps.py
@@ -5,22 +5,13 @@
 #
 # Licensed under the MIT License
 
-"""Tests for network classes.
+"""Tests for :class:`mqt.yaqs.core.data_structures.mps.MPS`."""
 
-This module provides unit tests for the Matrix Product MPS (MPS) class and its associated methods.
-It verifies correct initialization, custom tensor assignment, bond dimension computation, network flipping,
-orthogonality center shifting, normalization, observable measurement, and overall validity of MPS objects.
-These tests ensure that the MPS class functions as expected in various simulation scenarios.
-"""
-
-# ignore non-lowercase variable names for physics notation
 # ruff: noqa: N806
 
 from __future__ import annotations
 
 import copy
-import re
-from typing import TYPE_CHECKING
 
 import numpy as np
 import opt_einsum as oe
@@ -29,23 +20,17 @@ from qiskit.circuit import QuantumCircuit
 from scipy.stats import unitary_group
 
 from mqt.yaqs import simulator
+from mqt.yaqs.core.data_structures.mps import MPS
 from mqt.yaqs.core.data_structures.simulation_parameters import (
     AnalogSimParams,
+    Observable,
     StrongSimParams,
 )
-
-if TYPE_CHECKING:
-    from numpy.typing import NDArray
-
-from mqt.yaqs.core.data_structures.networks import MPO, MPS
-from mqt.yaqs.core.data_structures.simulation_parameters import Observable
 from mqt.yaqs.core.data_structures.state import State
 from mqt.yaqs.core.libraries.gate_library import BaseGate, Destroy, GateLibrary, Id, X, Z
 
-# ---- single-qubit ops ----
 _I2 = np.eye(2, dtype=complex)
 _X2 = np.array([[0, 1], [1, 0]], dtype=complex)
-_Y2 = np.array([[0, -1j], [1j, 0]], dtype=complex)
 _Z2 = np.array([[1, 0], [0, -1]], dtype=complex)
 
 
@@ -133,274 +118,6 @@ def _spin_current_bond_matrix(j_coupling: float) -> np.ndarray:
     y = np.array([[0.0, -1.0j], [1.0j, 0.0]], dtype=np.complex128)
     return 0.25 * j_coupling * (np.kron(x, y) - np.kron(y, x))
 
-
-def _embed_one_body(op: np.ndarray, length: int, i: int) -> np.ndarray:
-    """Embed a single-site operator into a length-L qubit Hilbert space.
-
-    Args:
-        op: Local 2x2 operator acting on site i.
-        length: Total number of sites.
-        i: Site index at which to apply the operator.
-
-    Returns:
-        Dense (2**length, 2**length) matrix representing I⊗…⊗op_i⊗…⊗I.
-    """
-    out = np.array([[1.0]], dtype=complex)
-    for k in range(length):
-        out = np.kron(out, op if k == i else _I2)
-    return out
-
-
-def _embed_two_body(op1: np.ndarray, op2: np.ndarray, length: int, i: int) -> np.ndarray:
-    """Embed a nearest-neighbor two-site operator into a length-L qubit Hilbert space.
-
-    Args:
-        op1: Local operator acting on site i.
-        op2: Local operator acting on site i+1.
-        length: Total number of sites.
-        i: Left site index of the two-body term.
-
-    Returns:
-        Dense (2**length, 2**length) matrix representing
-        I⊗…⊗op1_i⊗op2_{i+1}⊗…⊗I.
-    """
-    out = np.array([[1.0]], dtype=complex)
-    for k in range(length):
-        if k == i:
-            out = np.kron(out, op1)
-        elif k == i + 1:
-            out = np.kron(out, op2)
-        else:
-            out = np.kron(out, _I2)
-    return out
-
-
-def _ising_dense(length: int, j_val: float, g: float) -> np.ndarray:
-    """Construct the dense Ising Hamiltonian for an open chain.
-
-    The Hamiltonian is
-        H = -J sum_i Z_i Z_{i+1} - g sum_i X_i.
-
-    Args:
-        length: Number of sites.
-        j_val: Nearest-neighbor coupling strength.
-        g: Transverse-field strength.
-
-    Returns:
-        Dense (2**length, 2**length) Hamiltonian matrix.
-    """
-    dim = 2**length
-    H = np.zeros((dim, dim), dtype=complex)
-
-    for i in range(length - 1):
-        H += (-j_val) * _embed_two_body(_Z2, _Z2, length, i)
-    for i in range(length):
-        H += (-g) * _embed_one_body(_X2, length, i)
-
-    return H
-
-
-def _heisenberg_dense(length: int, jx: float, jy: float, jz: float, h: float) -> np.ndarray:
-    """Construct the dense Heisenberg Hamiltonian for an open chain.
-
-    The Hamiltonian is
-        H = -sum_i (Jx X_i X_{i+1} + Jy Y_i Y_{i+1} + Jz Z_i Z_{i+1}) - h sum_i Z_i.
-
-    Args:
-        length: Number of sites.
-        jx: XX coupling strength.
-        jy: YY coupling strength.
-        jz: ZZ coupling strength.
-        h: Longitudinal field strength.
-
-    Returns:
-        Dense (2**length, 2**length) Hamiltonian matrix.
-    """
-    dim = 2**length
-    H = np.zeros((dim, dim), dtype=complex)
-
-    for i in range(length - 1):
-        H += (-jx) * _embed_two_body(_X2, _X2, length, i)
-        H += (-jy) * _embed_two_body(_Y2, _Y2, length, i)
-        H += (-jz) * _embed_two_body(_Z2, _Z2, length, i)
-    for i in range(length):
-        H += (-h) * _embed_one_body(_Z2, length, i)
-
-    return H
-
-
-def _bose_hubbard_dense(length: int, local_dim: int, omega: float, hopping_j: float, hubbard_u: float) -> np.ndarray:
-    """Construct the exact dense Bose-Hubbard Hamiltonian for comparison.
-
-    Returns:
-        Dense Hamiltonian matrix.
-    """
-    # Local operators
-    a = Destroy(local_dim).matrix
-    adag = Destroy(local_dim).dag().matrix
-    n = adag @ a
-    id_op = np.eye(local_dim, dtype=complex)
-
-    dim = local_dim**length
-    H = np.zeros((dim, dim), dtype=complex)
-
-    # Build H term-by-term using Kronecker products
-    def embed(op_list: list[np.ndarray]) -> np.ndarray:
-        out = np.array([[1.0]], dtype=complex)
-        for op in op_list:
-            out = np.kron(out, op)
-        return out
-
-    # Onsite terms
-    for i in range(length):
-        op_list = [id_op] * length
-        op_list[i] = omega * n + 0.5 * hubbard_u * (n @ (n - id_op))
-        H += embed(op_list)
-
-    # Hopping terms
-    for i in range(length - 1):
-        # adag_i * a_{i+1}
-        op_list1 = [id_op] * length
-        op_list1[i] = adag
-        op_list1[i + 1] = a
-        H += -hopping_j * embed(op_list1)
-
-        # a_i * adag_{i+1}
-        op_list2 = [id_op] * length
-        op_list2[i] = a
-        op_list2[i + 1] = adag
-        H += -hopping_j * embed(op_list2)
-
-    return H
-
-
-def _embed_local_ops(length: int, local_dim: int, site_ops: list[np.ndarray]) -> np.ndarray:
-    """Embed a list of single-site operators into the full Hilbert space.
-
-    Returns:
-        np.ndarray: The embedded operator on the full chain Hilbert space.
-    """
-    identity = np.eye(local_dim, dtype=complex)
-    op_list = [identity] * length
-    for site, op in enumerate(site_ops):
-        op_list[site] = op
-    out = np.array([[1.0]], dtype=complex)
-    for op in op_list:
-        out = np.kron(out, op)
-    return out
-
-
-def _fermi_hubbard_1d_fermionic_dense(length: int, t: float, u: float) -> np.ndarray:
-    r"""Dense 1D Fermi-Hubbard Hamiltonian matching ``MPO.fermi_hubbard_1d``.
-
-    Uses open boundaries and
-    :math:`H = -t \\sum_{i,\\sigma} (c^\\dagger_{i,\\sigma} c_{i+1,\\sigma} + \\mathrm{h.c.})
-    + U \\sum_i n_{i,\\uparrow} n_{i,\\downarrow}` on local dimension-4 sites
-    (basis :math:`|0\\rangle, |\\!\\downarrow\\rangle, |\\!\\uparrow\\rangle, |\\!\\uparrow\\downarrow\\rangle`).
-
-    Returns:
-        np.ndarray: Dense Hamiltonian matrix of shape ``(4**length, 4**length)``.
-    """
-    local_dim = 4
-    c = np.array([[0, 1], [0, 0]], dtype=complex)
-    c_dag = np.array([[0, 0], [1, 0]], dtype=complex)
-    identity2 = np.eye(2, dtype=complex)
-    c_up = np.kron(c, identity2)
-    c_down = np.kron(identity2, c)
-    c_up_dag = np.kron(c_dag, identity2)
-    c_down_dag = np.kron(identity2, c_dag)
-    n_up = c_up_dag @ c_up
-    n_down = c_down_dag @ c_down
-    identity4 = np.eye(local_dim, dtype=complex)
-    onsite = u * n_up @ n_down
-
-    dim = local_dim**length
-    h = np.zeros((dim, dim), dtype=complex)
-
-    for site in range(length):
-        site_ops = [identity4] * length
-        site_ops[site] = onsite
-        h += _embed_local_ops(length, local_dim, site_ops)
-
-    for site in range(length - 1):
-        for c_right, c_left in ((c_up, c_up_dag), (c_down, c_down_dag)):
-            hop_ops = [identity4] * length
-            hop_ops[site] = c_left
-            hop_ops[site + 1] = c_right
-            h += -t * _embed_local_ops(length, local_dim, hop_ops)
-
-            hop_ops = [identity4] * length
-            hop_ops[site] = c_right
-            hop_ops[site + 1] = c_left
-            h += -t * _embed_local_ops(length, local_dim, hop_ops)
-
-    return h
-
-
-def _fermi_hubbard_1d_jordan_wigner_dense(num_orbitals: int, t: float, u: float) -> np.ndarray:
-    """Dense JW-transformed Fermi-Hubbard on an interleaved spin chain.
-
-    ``num_orbitals`` must be even. Orbitals are ordered 1↑, 1↓, 2↑, 2↓, ...
-    and the Hamiltonian matches the docstring of ``MPO.fermi_hubbard_1d(..., jordan_wigner=True)``.
-
-    Returns:
-        np.ndarray: Dense Hamiltonian matrix of shape ``(2**num_orbitals, 2**num_orbitals)``.
-
-    Raises:
-        ValueError: If ``num_orbitals`` is odd.
-    """
-    if num_orbitals % 2 != 0:
-        msg = "num_orbitals must be even."
-        raise ValueError(msg)
-
-    num_sites = num_orbitals // 2
-    dim = 2**num_orbitals
-    h = np.zeros((dim, dim), dtype=complex)
-
-    def term(ops: list[tuple[int, np.ndarray]]) -> np.ndarray:
-        local = [_I2] * num_orbitals
-        for index, op in ops:
-            local[index] = op
-        return _embed_local_ops(num_orbitals, 2, local)
-
-    for site in range(num_sites):
-        up = 2 * site
-        down = 2 * site + 1
-        h += (u / 4) * term([(up, _I2), (down, _I2)])
-        h += -(u / 4) * term([(up, _Z2)])
-        h += -(u / 4) * term([(down, _Z2)])
-        h += (u / 4) * term([(up, _Z2), (down, _Z2)])
-
-    for site in range(num_sites - 1):
-        up = 2 * site
-        down = 2 * site + 1
-        up_next = 2 * (site + 1)
-        h += -(t / 2) * term([(up, _X2), (down, _Z2), (up_next, _X2)])
-        h += -(t / 2) * term([(up, _Y2), (down, _Z2), (up_next, _Y2)])
-
-        down_next = 2 * (site + 1) + 1
-        h += -(t / 2) * term([(down, _X2), (up_next, _Z2), (down_next, _X2)])
-        h += -(t / 2) * term([(down, _Y2), (up_next, _Z2), (down_next, _Y2)])
-
-    return h
-
-
-def untranspose_block(mpo_tensor: NDArray[np.complex128]) -> NDArray[np.complex128]:
-    """Reverse the transposition of an MPO tensor.
-
-    MPO tensors are stored in the order (sigma, sigma', row, col). This function transposes
-    the tensor to the order (row, col, sigma, sigma') so that the first two indices can be interpreted
-    as a block matrix of operators.
-
-    Args:
-        mpo_tensor (NDArray[np.complex128]): The MPO tensor in (sigma, sigma', row, col) order.
-
-    Returns:
-        NDArray[np.complex128]: The MPO tensor in (row, col, sigma, sigma') order.
-    """
-    return np.transpose(mpo_tensor, (2, 3, 0, 1))
-
-
 def crandn(
     size: int | tuple[int, ...],
     *args: int,
@@ -460,343 +177,7 @@ def _expected_uniform_clipped_bonds(length: int, chi_max: int) -> list[int]:
     bonds.extend(min(chi_max, 2 ** min(i, length - i)) for i in range(1, length))
     bonds.append(1)
     return bonds
-
-
 rng = np.random.default_rng()
-
-##############################################################################
-# Tests for the MPO class
-##############################################################################
-
-
-def test_ising_correct_operator() -> None:
-    """Verify that the Ising MPO matches the exact dense Hamiltonian."""
-    L = 5
-    J = 1.0
-    g = 0.5
-
-    mpo = MPO.ising(L, J, g)
-
-    assert mpo.length == L
-    assert mpo.physical_dimension == 2
-    assert len(mpo.tensors) == L
-
-    assert np.allclose(mpo.to_matrix(), _ising_dense(L, J, g), atol=1e-12)
-
-
-def test_heisenberg_correct_operator() -> None:
-    """Verify that the Heisenberg MPO matches the exact dense Hamiltonian."""
-    L = 5
-    Jx, Jy, Jz, h = 1.0, 0.5, 0.3, 0.2
-
-    mpo = MPO.heisenberg(L, Jx, Jy, Jz, h)
-
-    assert np.allclose(mpo.to_matrix(), _heisenberg_dense(L, Jx, Jy, Jz, h), atol=1e-12)
-
-
-def test_bose_hubbard_correct_operator() -> None:
-    """Verify that the Bose-Hubbard MPO matches the exact dense Hamiltonian."""
-    length = 4
-    local_dim = 3  # up to 2 bosons per site
-    omega = 0.7
-    J = 0.2
-    U = 1.3
-
-    mpo = MPO.bose_hubbard(
-        length=length,
-        local_dim=local_dim,
-        omega=omega,
-        hopping_j=J,
-        hubbard_u=U,
-    )
-
-    # Basic checks
-    assert mpo.length == length
-    assert mpo.physical_dimension == local_dim
-    assert len(mpo.tensors) == length
-    assert all(t.shape[2] <= 4 and t.shape[3] <= 4 for t in mpo.tensors), "Bond dimension should be 4"
-
-    # Dense comparison
-    H_dense = _bose_hubbard_dense(length, local_dim, omega, J, U)
-    H_mpo = mpo.to_matrix()
-    np.testing.assert_allclose(H_mpo, H_dense, atol=1e-8)
-
-
-def test_fermi_hubbard_1d_correct_operator() -> None:
-    """Verify the fermionic 1D Fermi-Hubbard MPO matches the dense Hamiltonian."""
-    length = 3
-    u, t = 0.5, 1.0
-
-    mpo = MPO.fermi_hubbard_1d(length, t, u)
-
-    assert mpo.length == length
-    assert mpo.physical_dimension == 4
-    assert len(mpo.tensors) == length
-    assert all(tensor.shape[2] <= 6 and tensor.shape[3] <= 6 for tensor in mpo.tensors)
-
-    h_dense = _fermi_hubbard_1d_fermionic_dense(length, t, u)
-    np.testing.assert_allclose(mpo.to_matrix(), h_dense, atol=1e-10)
-
-
-def test_fermi_hubbard_1d_jordan_wigner_correct_operator() -> None:
-    """Verify the JW 1D Fermi-Hubbard MPO matches the dense Pauli Hamiltonian."""
-    num_orbitals = 4
-    u, t = 0.5, 1.0
-
-    mpo = MPO.fermi_hubbard_1d(num_orbitals, t, u, jordan_wigner=True)
-
-    assert mpo.length == num_orbitals
-    assert mpo.physical_dimension == 2
-    assert len(mpo.tensors) == num_orbitals
-    assert all(tensor.shape[2] <= 16 and tensor.shape[3] <= 16 for tensor in mpo.tensors)
-
-    h_dense = _fermi_hubbard_1d_jordan_wigner_dense(num_orbitals, t, u)
-    np.testing.assert_allclose(mpo.to_matrix(), h_dense, atol=1e-10)
-
-    with pytest.raises(ValueError, match=re.escape("length must be an even integer ≥ 2 (ordering: 1↑,1↓,2↑,2↓,...).")):
-        MPO.fermi_hubbard_1d(length=5, t=t, u=u, jordan_wigner=True)
-
-
-def test_fermi_hubbard_1d_length_one() -> None:
-    """Verify a single fermionic site MPO matches the dense reference."""
-    length = 1
-    u, t = 0.5, 1.0
-
-    mpo = MPO.fermi_hubbard_1d(length, t, u)
-
-    assert mpo.length == length
-    assert mpo.physical_dimension == 4
-    assert mpo.tensors[0].shape == (4, 4, 1, 1)
-
-    h_dense = _fermi_hubbard_1d_fermionic_dense(length, t, u)
-    np.testing.assert_allclose(mpo.to_matrix(), h_dense, atol=1e-10)
-
-
-def test_fermi_hubbard_1d_cross_representation() -> None:
-    """Onsite terms agree between fermionic and JW MPOs under the site basis map.
-
-      Hopping terms differ between representations (composite fermionic sites vs JW
-    qubit chain), so this test uses ``t=0`` to compare only the interaction part.
-    """
-    u = 0.5
-    for length in (1, 2, 3):
-        h_ferm = MPO.fermi_hubbard_1d(length, t=0.0, u=u).to_matrix()
-        h_jw = MPO.fermi_hubbard_1d(2 * length, t=0.0, u=u, jordan_wigner=True).to_matrix()
-        np.testing.assert_allclose(h_ferm, h_jw, atol=1e-10)
-
-
-def test_identity() -> None:
-    """Test that identity initializes an identity MPO correctly.
-
-    This test checks that an identity MPO has the correct length, physical dimension,
-    and that each tensor corresponds to the identity operator.
-    """
-    mpo = MPO()
-    length = 3
-    pdim = 2
-
-    mpo.identity(length, physical_dimension=pdim)
-
-    assert mpo.length == length
-    assert mpo.physical_dimension == pdim
-    assert len(mpo.tensors) == length
-
-    for tensor in mpo.tensors:
-        assert tensor.shape == (2, 2, 1, 1)
-        assert np.allclose(np.squeeze(tensor), Id().matrix)
-
-
-def test_finite_state_machine() -> None:
-    """Test initializing a custom Hamiltonian MPO using user-provided boundary and inner tensors.
-
-    This test creates random tensors for the left boundary, inner sites, and right boundary,
-    initializes the MPO with these using finite_state_machine, and verifies that the tensors
-    have the expected shapes and values (after appropriate transposition).
-    """
-    length = 4
-    pdim = 2
-
-    left_bound = rng.random(size=(1, 2, pdim, pdim)).astype(np.complex128)
-    inner = rng.random(size=(2, 2, pdim, pdim)).astype(np.complex128)
-    right_bound = rng.random(size=(2, 1, pdim, pdim)).astype(np.complex128)
-
-    mpo = MPO()
-    mpo.finite_state_machine(length, left_bound, inner, right_bound)
-
-    assert mpo.length == length
-    assert len(mpo.tensors) == length
-
-    assert mpo.tensors[0].shape == (pdim, pdim, 1, 2)
-    for i in range(1, length - 1):
-        assert mpo.tensors[i].shape == (pdim, pdim, 2, 2)
-    assert mpo.tensors[-1].shape == (pdim, pdim, 2, 1)
-
-    assert np.allclose(mpo.tensors[0], np.transpose(left_bound, (2, 3, 0, 1)))
-    for i in range(1, length - 1):
-        assert np.allclose(mpo.tensors[i], np.transpose(inner, (2, 3, 0, 1)))
-    assert np.allclose(mpo.tensors[-1], np.transpose(right_bound, (2, 3, 0, 1)))
-
-
-def test_custom() -> None:
-    """Test that custom correctly sets up an MPO from a user-provided list of tensors.
-
-    This test provides a list of tensors for the left boundary, middle, and right boundary,
-    initializes the MPO, and checks that the shapes and values of the MPO tensors match the inputs.
-    """
-    length = 3
-    pdim = 2
-    tensors = [
-        rng.random(size=(1, 2, pdim, pdim)).astype(np.complex128),
-        rng.random(size=(2, 2, pdim, pdim)).astype(np.complex128),
-        rng.random(size=(2, 1, pdim, pdim)).astype(np.complex128),
-    ]
-
-    mpo = MPO()
-    mpo.custom(tensors)
-
-    assert mpo.length == length
-    assert mpo.physical_dimension == pdim
-    assert len(mpo.tensors) == length
-
-    for original, created in zip(tensors, mpo.tensors, strict=False):
-        assert original.shape == created.shape
-        assert np.allclose(original, created)
-
-
-def test_from_matrix() -> None:
-    """Test that from_matrix() constructs a correct MPO.
-
-    This test constructs a dense Bose-Hubbard Hamiltonian and creates an MPO via from_matrix().
-    It checks:
-    - reconstruction correctness for Bose-Hubbard
-    - random matrices at very large bond dimension
-    - random matrices at moderately truncated bond dimension
-    - all validation error branches (Codecov)
-    """
-    rng = np.random.default_rng()
-
-    length = 5
-    d = 3  # local dimension
-    H = _bose_hubbard_dense(length, d, 0.9, 0.6, 0.2)
-
-    Hmpo = MPO.from_matrix(H, d, 4)
-    assert np.allclose(H, Hmpo.to_matrix())
-
-    H = rng.random((d**length, d**length)) + 1j * rng.random((d**length, d**length))
-    Hmpo = MPO.from_matrix(H, d, 1_000_000)
-    assert np.allclose(H, Hmpo.to_matrix())
-
-    length = 6
-    H = rng.random((d**length, d**length)) + 1j * rng.random((d**length, d**length))
-    Hmpo = MPO.from_matrix(H, d, 728)
-    assert np.max(np.abs(H - Hmpo.to_matrix())) < 1e-2
-
-    mat = np.eye(1)
-    with pytest.raises(ValueError, match="Physical dimension d must be > 0"):
-        MPO.from_matrix(mat, d=0)
-
-    # non-square matrix
-    mat = np.zeros((4, 2))
-    with pytest.raises(ValueError, match="Matrix must be square"):
-        MPO.from_matrix(mat, d=2)
-
-    # d == 1 but matrix not 1x1
-    mat = np.eye(4)
-    with pytest.raises(ValueError, match="1x1"):
-        MPO.from_matrix(mat, d=1)
-
-    # matrix dimension not a power of d
-    mat = np.eye(6)
-    with pytest.raises(ValueError, match="not a power"):
-        MPO.from_matrix(mat, d=2)
-
-    # inferred n < 1 (log(1)/log(100) = 0)
-    mat = np.eye(1)
-    with pytest.raises(ValueError, match="invalid"):
-        MPO.from_matrix(mat, d=100)
-
-
-def test_to_mps() -> None:
-    """Test converting an MPO to an MPS.
-
-    This test initializes an MPO using ising, converts it to an MPS via to_mps,
-    and verifies that the resulting MPS has the correct length and that each tensor has been reshaped
-    to the expected dimensions.
-    """
-    length = 3
-    J, g = 1.0, 0.5
-
-    mpo = MPO.ising(length, J, g)
-    mps = mpo.to_mps()
-
-    assert isinstance(mps, MPS)
-    assert mps.length == length
-
-    for i, tensor in enumerate(mps.tensors):
-        original_mpo_tensor = mpo.tensors[i]
-        pdim2 = original_mpo_tensor.shape[0] * original_mpo_tensor.shape[1]
-        bond_in = original_mpo_tensor.shape[2]
-        bond_out = original_mpo_tensor.shape[3]
-        assert tensor.shape == (pdim2, bond_in, bond_out)
-
-
-def test_check_if_valid_mpo() -> None:
-    """Test that a valid MPO passes the check_if_valid_mpo method without raising errors.
-
-    This test initializes an Ising MPO and calls check_if_valid_mpo, which should validate the MPO.
-    """
-    length = 4
-    J, g = 1.0, 0.5
-
-    mpo = MPO.ising(length, J, g)
-    mpo.check_if_valid_mpo()
-
-
-def test_rotate() -> None:
-    """Test the rotate method for an MPO.
-
-    This test checks that rotating an MPO (without conjugation) transposes each tensor as expected,
-    and that rotating back with conjugation returns tensors with the original physical dimensions.
-    """
-    length = 3
-    J, g = 1.0, 0.5
-
-    mpo = MPO.ising(length, J, g)
-    original_tensors = [t.copy() for t in mpo.tensors]
-
-    mpo.rotate(conjugate=False)
-    for orig, rotated in zip(original_tensors, mpo.tensors, strict=False):
-        assert rotated.shape == (
-            orig.shape[1],
-            orig.shape[0],
-            orig.shape[2],
-            orig.shape[3],
-        )
-        np.testing.assert_allclose(rotated, np.transpose(orig, (1, 0, 2, 3)))
-
-    mpo.rotate(conjugate=True)
-    for tensor in mpo.tensors:
-        assert tensor.shape[0:2] == (2, 2)
-
-
-def test_check_if_identity() -> None:
-    """Test that an identity MPO is recognized as identity by check_if_identity.
-
-    This test initializes an identity MPO and verifies that check_if_identity returns True
-    when a fidelity threshold is provided.
-    """
-    mpo = MPO()
-    length = 3
-    pdim = 2
-
-    mpo.identity(length, pdim)
-    fidelity_threshold = 0.9
-    assert mpo.check_if_identity(fidelity_threshold) is True
-
-
-##############################################################################
-# Tests for the MPS class
-##############################################################################
 
 
 @pytest.mark.parametrize("state", ["zeros", "ones", "x+", "x-", "y+", "y-", "Neel", "wall", "basis"])
@@ -1884,187 +1265,3 @@ def test_evaluate_observables_meta_validation_errors() -> None:
         mps.evaluate_observables(sim_non_adj, results_adj, column_index=0)
 
 
-def test_hamiltonian_raises_on_nonpositive_length() -> None:
-    """Hamiltonian input validation: non-positive system size must raise."""
-    with pytest.raises(ValueError, match=r"L must be positive\."):
-        MPO.hamiltonian(length=0)
-
-    with pytest.raises(ValueError, match=r"L must be positive\."):
-        MPO.hamiltonian(length=-3)
-
-
-def test_hamiltonian_raises_on_invalid_bc() -> None:
-    """Hamiltonian input validation: unsupported boundary conditions must raise."""
-    with pytest.raises(ValueError, match=r"bc must be 'open' or 'periodic'\."):
-        MPO.hamiltonian(length=4, bc="closed")
-
-    with pytest.raises(ValueError, match=r"bc must be 'open' or 'periodic'\."):
-        MPO.hamiltonian(length=4, bc="")
-
-
-def test_hamiltonian_raises_on_invalid_one_body_operator() -> None:
-    """Hamiltonian input validation: invalid single-site operator labels must raise."""
-    with pytest.raises(ValueError, match=r"Invalid operator 'Q'"):
-        MPO.hamiltonian(length=3, one_body=[(1.0, "Q")])
-
-
-def test_hamiltonian_raises_on_invalid_two_body_operator_left() -> None:
-    """Hamiltonian input validation: invalid left two-body operator labels must raise."""
-    with pytest.raises(ValueError, match=r"Invalid operator 'Q'"):
-        MPO.hamiltonian(length=3, two_body=[(1.0, "Q", "Z")])
-
-
-def test_hamiltonian_raises_on_invalid_two_body_operator_right() -> None:
-    """Hamiltonian input validation: invalid right two-body operator labels must raise."""
-    with pytest.raises(ValueError, match=r"Invalid operator 'Q'"):
-        MPO.hamiltonian(length=3, two_body=[(1.0, "X", "Q")])
-
-
-def test_hamiltonian_normalizes_operator_case() -> None:
-    """Hamiltonian construction: operator labels are case-insensitive and normalized."""
-    _ = MPO.hamiltonian(
-        length=2,
-        one_body=[(0.5, "x")],
-        two_body=[(1.0, "z", "y")],
-        bc="open",
-        n_sweeps=0,
-    )
-
-
-def test_from_pauli_sum_raises_on_invalid_physical_dimension() -> None:
-    """Pauli-sum MPO validation: only physical_dimension=2 is supported."""
-    mpo = MPO()
-    with pytest.raises(ValueError, match=r"Only physical_dimension=2 is supported"):
-        mpo.from_pauli_sum(terms=[(1.0, "Z0")], length=2, physical_dimension=3)
-
-
-def test_from_pauli_sum_raises_on_nonpositive_length() -> None:
-    """Pauli-sum MPO validation: non-positive length must raise."""
-    mpo = MPO()
-    with pytest.raises(ValueError, match=r"length must be positive\."):
-        mpo.from_pauli_sum(terms=[(1.0, "Z0")], length=0)
-
-    with pytest.raises(ValueError, match=r"length must be positive\."):
-        mpo.from_pauli_sum(terms=[(1.0, "Z0")], length=-5)
-
-
-def test_from_pauli_sum_raises_on_site_index_out_of_bounds(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Pauli-sum MPO validation: parsed site indices outside [0, L-1] must raise."""
-    mpo = MPO()
-
-    # Force the parser to return an out-of-bounds site index regardless of spec.
-    monkeypatch.setattr(mpo, "_parse_pauli_string", lambda _spec: {99: "Z"})
-
-    with pytest.raises(ValueError, match=r"Site index 99 outside \[0, 3\]\."):
-        mpo.from_pauli_sum(terms=[(1.0, "Z0")], length=4)
-
-
-def test_from_pauli_sum_raises_on_invalid_local_op_label(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Pauli-sum MPO validation: parsed local operator labels must be in _VALID."""
-    mpo = MPO()
-
-    # Force the parser to return an invalid label.
-    monkeypatch.setattr(mpo, "_parse_pauli_string", lambda _spec: {0: "Q"})
-
-    with pytest.raises(ValueError, match=r"Invalid local op 'Q'"):
-        mpo.from_pauli_sum(terms=[(1.0, "Z0")], length=2)
-
-
-def test_from_pauli_sum_empty_terms_builds_zero_mpo() -> None:
-    """Pauli-sum MPO construction: empty term list yields an all-zero MPO with bond dim 1."""
-    mpo = MPO()
-    mpo.from_pauli_sum(terms=[], length=3, n_sweeps=0)  # n_sweeps=0 keeps it fast
-
-    assert len(mpo.tensors) == 3
-    for t in mpo.tensors:
-        assert t.shape == (2, 2, 1, 1)
-        assert np.allclose(t, 0.0)
-
-
-def test_compress_raises_on_negative_n_sweeps() -> None:
-    """MPO compress input validation: negative n_sweeps must raise."""
-    mpo = MPO()
-    mpo.tensors = [np.zeros((2, 2, 1, 1), dtype=complex)]
-    with pytest.raises(ValueError, match=r"n_sweeps must be >= 0\."):
-        mpo.compress(n_sweeps=-1)
-
-
-def test_compress_raises_on_invalid_directions() -> None:
-    """MPO compress input validation: invalid sweep schedule strings must raise."""
-    mpo = MPO()
-    mpo.tensors = [np.zeros((2, 2, 1, 1), dtype=complex)]
-    with pytest.raises(
-        ValueError,
-        match=r"directions must be one of \{'lr', 'rl', 'lr_rl', 'rl_lr'\}\.",
-    ):
-        mpo.compress(directions="lr,rl")
-
-
-def test_compress_n_sweeps_zero_returns_without_calling_sweeps(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """MPO compress control flow: n_sweeps=0 must return without invoking sweeps."""
-    mpo = MPO()
-    mpo.tensors = [
-        np.zeros((2, 2, 1, 1), dtype=complex),
-        np.zeros((2, 2, 1, 1), dtype=complex),
-    ]
-
-    called = False
-
-    def boom(**_kwargs: object) -> None:
-        nonlocal called
-        called = True
-        msg = "should not be called when n_sweeps=0"
-        raise AssertionError(msg)
-
-    monkeypatch.setattr(mpo, "_compress_one_sweep", boom)
-
-    mpo.compress(n_sweeps=0, directions="lr_rl")
-    assert called is False
-
-
-def test_compress_one_sweep_raises_on_invalid_direction() -> None:
-    """MPO _compress_one_sweep input validation: direction must be 'lr' or 'rl'."""
-    mpo = MPO()
-    mpo.tensors = [
-        np.zeros((2, 2, 1, 1), dtype=complex),
-        np.zeros((2, 2, 1, 1), dtype=complex),
-    ]
-    with pytest.raises(ValueError, match=r"direction must be 'lr' or 'rl'\."):
-        mpo._compress_one_sweep(direction="xx", tol=1e-12, max_bond_dim=None)  # noqa: SLF001
-
-
-def test_from_pauli_sum_empty_spec_is_identity_term() -> None:
-    """Pauli parsing integration: empty spec denotes the identity operator."""
-    mpo = MPO()
-    mpo.from_pauli_sum(terms=[(1.0, "")], length=2, n_sweeps=0)
-    assert len(mpo.tensors) == 2  # construction succeeded
-
-
-def test_from_pauli_sum_parses_commas_and_normalizes_case() -> None:
-    """Pauli parsing integration: commas/whitespace are accepted and labels are case-normalized."""
-    mpo = MPO()
-    mpo.from_pauli_sum(terms=[(1.0, "x0, y1")], length=2, n_sweeps=0)
-    assert len(mpo.tensors) == 2
-
-
-def test_from_pauli_sum_raises_on_duplicate_site_in_spec() -> None:
-    """Pauli parsing integration: duplicate site indices in a spec must raise."""
-    mpo = MPO()
-    with pytest.raises(ValueError, match=r"Duplicate site 0 in spec"):
-        mpo.from_pauli_sum(terms=[(1.0, "X0 Z0")], length=2, n_sweeps=0)
-
-
-def test_from_pauli_sum_raises_on_invalid_tokens_in_spec() -> None:
-    """Pauli parsing integration: invalid tokens in the spec must raise."""
-    mpo = MPO()
-    with pytest.raises(ValueError, match=r"Invalid token\(s\) in spec"):
-        mpo.from_pauli_sum(terms=[(1.0, "X0 Q2")], length=3, n_sweeps=0)
-
-    with pytest.raises(ValueError, match=r"Invalid token\(s\) in spec"):
-        mpo.from_pauli_sum(terms=[(1.0, "X0 Y2 garbage")], length=4, n_sweeps=0)
